pyproject.toml
[project]
name = "maze"
version = "0.1.0"
description = ""
authors = [
    {name = "Giacomo De Florio", email = "giacomo.deflorio.dev@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pygame-ce (>=2.5.5,<3.0.0)",
]

[tool.poetry]
packages = [{include = "maze", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
